+++ back/src/main/java/com/thered/stocksignal/jwt/JWTFilter.java
@@ -26,7 +26,6 @@ public class JWTFilter extends OncePerRequestFilter {
         String authorization = request.getHeader("Authorization");
 
         if(authorization == null || !authorization.startsWith("Bearer ")){
-            System.out.println("token null");
             filterChain.doFilter(request, response);
             return; //조건이 해당되면 메소드 종료(필수)
         }
