+++ back/src/main/java/com/thered/stocksignal/websocket/WebSocketHandler.java
@@ -17,6 +17,7 @@ import com.thered.stocksignal.service.user.UserAccountService;
 import jakarta.annotation.PostConstruct;
 import jakarta.websocket.Session;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.CloseStatus;
 import org.springframework.web.socket.TextMessage;
@@ -29,6 +30,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
+@Slf4j
 @Component
 @RequiredArgsConstructor
 public class WebSocketHandler extends TextWebSocketHandler {
@@ -48,12 +50,12 @@ public class WebSocketHandler extends TextWebSocketHandler {
 
     @Override
     public void afterConnectionEstablished(WebSocketSession session) throws Exception {
-        System.out.println("새로운 WebSocket 세션이 연결되었습니다: " + session.getId());
+        log.info("새로운 WebSocket 세션이 연결되었습니다: {}", session.getId());
     }
 
     @Override
     public void afterConnectionClosed(WebSocketSession session, CloseStatus status) throws Exception {
-        System.out.println("WebSocket 세션이 종료되었습니다: " + session.getId());
+        log.info("WebSocket 세션이 종료되었습니다:  {}", session.getId());
     }
 
     @Override
@@ -72,11 +74,9 @@ public class WebSocketHandler extends TextWebSocketHandler {
             String companyCode = companyService.findCodeByName(companyName).get().getCompanyCode();
 
             if (userId == -1) {
-                System.out.println("유효하지 않은 세션입니다. 유저 확인 불가능");
+                throw new IllegalArgumentException("유효하지 않은 세션입니다. 유저 확인 불가능");
             }
             else{
-                System.out.println("받은 메시지: " + payload);
-
                 if(userSessions.get(userId) == null){
                     userSessions.put(userId, new UserSession(clientSession, null));
                 }
@@ -93,7 +93,7 @@ public class WebSocketHandler extends TextWebSocketHandler {
                         handleKisSocketRequest(token, userId, companyCode, companyName, "2");
                         break;
                     default:
-                        System.out.println("알 수 없는 action: " + dto.getAction());
+                        throw new IllegalArgumentException("알 수 없는 action: " + dto.getAction());
                 }
             }
         } finally {
@@ -106,7 +106,7 @@ public class WebSocketHandler extends TextWebSocketHandler {
         String token =  (String)kisSession.getUserProperties().get("token");
         Long userId = userAccountService.getUserIdFromToken(token);// 토큰으로부터 사용자 ID 반환
         if (userId == -1) {
-            System.out.println("유효하지 않은 세션입니다. 유저 확인 불가능");
+            throw new IllegalArgumentException("유효하지 않은 세션입니다. 유저 확인 불가능");
         }
         else{
             UserSession userSession = userSessions.get(userId);
@@ -196,7 +196,7 @@ public class WebSocketHandler extends TextWebSocketHandler {
 
             // 기존 세션으로 업데이트된 메시지 전송
             existingSession.getKisSession().getAsyncRemote().sendText(message);
-            System.out.println("변경된 메시지를 전송했습니다: " + message);
+            log.info("변경된 메시지를 전송했습니다: {}", message);
 
             if(type.equals("2")){
                 // 삭제 후에 scenario 테이블에 아무것도 없다면 userSessions에서 세션을 가져오고 종료
@@ -228,7 +228,7 @@ public class WebSocketHandler extends TextWebSocketHandler {
                 scenarioTrade.checkAutoTrade(userId, dto);
             }
         } else {
-            System.out.println(response); // 초당 거래건수 초과 등
+            throw new RuntimeException("한투에서 온 응답이 예상된 방식과 다릅니다 : " + response); // 초당 거래건수 초과 등
         }
     }
 
@@ -248,7 +248,6 @@ public class WebSocketHandler extends TextWebSocketHandler {
 
         // 각 종목 정보 세트를 순회
         for (List<String> stockInfo : stockInfos) {
-            System.out.println(stockInfo);
 
             if (!stockInfo.isEmpty()) {
                 String companyCode = stockInfo.getFirst();
